backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
torch==2.3.0
torchvision==0.18.0
timm==0.9.16
Pillow==10.3.0
numpy==1.26.4
sqlalchemy==2.0.30
pyodbc==5.1.0
python-dotenv==1.0.1
google-generativeai==0.7.0
pydantic==2.7.1
pydantic-settings==2.3.0
httpx==0.27.0
